This is the first major update to the cli-spectrogram
We've transitioned away from a single threading application to a multi threading one
where the GUI logic happens independant of the background processes.
This speeds up rendering and handles user input in a more robust way.

This version has all the features and more included in the original cli-spectrogram.

TODO:
- Add auto config for DAQ driver json
- Add auto NFFT to fit legend and plot in term with hot key
- Add menu to change directories or params passed in original call
- Possibly make a mini legend mode for a more compact legend?